# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Node.js API with JWT Authentication",
    "description": "API for managing Users, Roles, Claims, and TestObjects with JWT-based Authentication"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      }
    },
    "schemas": {
...
      "TestObject": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Test Object 1" },
          "description": { "type": "string", "example": "Test Object 1 description" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Unauthorized" },
          "status": { "type": "integer", "example": 401 }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/testobjects": {
      "get": {
        "tags": ["TestObjects"],
        "summary": "Get all test objects",
        "description": "Retrieve a list of all test objects (User only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of test objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TestObject" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "post": {
        "tags": ["TestObjects"],
        "summary": "Create a test object",
        "description": "Create a new test object (User only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Test object created successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      }
    },
    "/testobjects/{id}": {
      "get": {
        "tags": ["TestObjects"],
        "summary": "Get a TestObject by ID",
        "description": "Retrieve a TestObjects by its ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "TestObject ID"
          }
        ],
        "responses": {
          "200": {
            "description": "TestObject retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TestObject" }
              }
            }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "TestObject not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "put": {
        "tags": ["TestObjects"],
        "summary": "Update TestObject by ID",
        "description": "Update a TestObject by its ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "TestObject ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "TestObject updated successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "TestObject not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "delete": {
        "tags": ["TestObjects"],
        "summary": "Delete a TestObject by ID",
        "description": "Delete a TestObject by its ID (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "TestObject ID"
          }
        ],
        "responses": {
          "200": { "description": "TestObject deleted successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "TestObject not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      }
    }
  }
}